Made some changes to the way daemon will call other modules
Moved the logic for DNS and database routines outside of the IP
header handler.  This means it requires some extra malloc()s to
bring the data back out to the packet handler function, but will
help decoupling the modules: As much as possible, all the calls
to other modules should happen in the packet handler.